Move manifest models/loaders to Helpers
Relocated package/editor manifest types and loader/deserializer classes from `Celbridge.Packages` under `Services` to `Celbridge.Packages.Helpers`, updating namespaces and file paths to better reflect their helper role. Updated `PackageRegistry` and related tests to import the new namespace.
